# Improve performance of COPY TO