--- a/NEWS.carto.md
+++ b/NEWS.carto.md
@@ -1,5 +1,11 @@
 # CARTO's Changelog
 
+## v1.2.0-carto.1
+Released 2018-mm-dd
+
+Bug fixes:
+ * Improves performance of COPY TO by sending bigger chunks through low level `push()`. See https://github.com/CartoDB/node-pg-copy-streams/pull/1
+
 ## v1.2.0
 Released 2016-08-22