@rafatower 2018-06-08
PS: it passes the tests (which were very useful BTW):

```
$ PGUSER=postgres PGDATABASE=test npm test 

> pg-copy-streams@1.2.0 test /home/rtorre_unencrypted/src/node-pg-copy-streams
> node test
```