@jeromew 2019-03-14
for info you should be able to use upstream node-pg-copy-streams version 2.0.0 if you prefer. It is basically the previous version 1.2.0 + your patches added. Sorry for the delay and inconvenience on this.